requirements.txt
numpy==1.19.5
scipy==1.5.4
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
opt-einsum==3.3.0
protobuf==3.14.0
PyYAML==5.4.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
wrapt==1.12.1